requirements.txt
streamlit
ultralytics
pillow
folium
streamlit-folium
pandas
numpy
boto3
opencv-python-headless
